refactor(iroha_core, iroha_data_model): move `trait Registrable` to the data model
There are functions such as `Account:new`, which create not `Account`,
but `NewAccount`. Sometimes it is not desirable and I, as a user of
`iroha_data_model` crate, want to simply create an `Account` itself.
The only way to do so without `transparent_api` feature is to use
`Registrable` trait (transforming e.g. `NewAccount` to `Account`), which
is for some reason defined in `iroha_core`.

This commit moves `Registrable` from `iroha_core` to `iroha_data_model`.

Signed-off-by: quacumque <quacumque@fastmail.jp>